[MRM-1473] remove use of plexus-spring fix module metadata-store-file
git-svn-id: https://svn.apache.org/repos/asf/archiva/trunk@1130154 13f79535-47bb-0310-9956-ffa450edef68

+++ archiva-modules/plugins/metadata-store-file/src/test/java/org/apache/archiva/metadata/repository/file/FileMetadataRepositoryTest.java
@@ -41,9 +41,11 @@ public class FileMetadataRepositoryTest
     {
         super.setUp();
 
-        File directory = getTestFile( "target/test-repositories" );
-        FileUtils.deleteDirectory( directory );
-
+        File directory = new File( "target/test-repositories" );
+        if (directory.exists())
+        {
+            FileUtils.deleteDirectory( directory );
+        }
         ArchivaConfiguration config = createTestConfiguration( directory );
         Map<String, MetadataFacetFactory> factories = createTestMetadataFacetFactories();
